refactor: extract cache module from main.rs
